Move node into separate crate
