+++ node/res/chain-spec-raw-gemini-3d.json
